pyproject.toml
[project]
name = "atelierlstm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mysql-connector-python>=9.0.0",
    "psycopg>=3.3.2",
    "python-dotenv>=1.0.1",
]
